Refactor rate service
Now fetcher is a struct not a function, with name property.
Split rate specific code into separate package. Add new currency
provider (PrivatBank API). Allow client's code to pass context to
GetRate() method.
